+++ Octave/dh_to_mat4_sintaylor.m
@@ -12,8 +12,8 @@ function links = dh_to_mat4_sintaylor(d,a,alpha, sin_output_order)
     for i = 1:n
         m = [
             ONE_FIXED 0 0 a(i);
-            0 cos_Nb64_taylor(alpha(i), sin_output_order) -sin_Nb64_taylor(alpha(i),sin_output_order) 0;
-            0 sin_Nb64_taylor(alpha(i),sin_output_order) cos_Nb64_taylor(alpha(i),sin_output_order) d(i);
+            0 int32(cos_Nb64_taylor(alpha(i), sin_output_order)) int32(-sin_Nb64_taylor(alpha(i),sin_output_order)) 0;
+            0 int32(sin_Nb64_taylor(alpha(i),sin_output_order)) int32(cos_Nb64_taylor(alpha(i),sin_output_order)) d(i);
             0 0 0 ONE_FIXED;
         ];
         links{i} = int32(m);
